Meetups/M13/requirements.txt
click==7.1.2
cycler==0.10.0
Cython==0.29.20
Flask==1.0.2
flask-rabbitmq-cv==0.0.3
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
marshmallow==3.6.1
matplotlib==3.2.1
numpy==1.18.5
pika==0.13.0
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
webargs==6.1.0
Werkzeug==0.15.6
